pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "data-art"
version = "0.1.0"
description = "Transform text into visual data art by encoding characters as RGB pixels"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pillow>=12.1.1",
]

[project.scripts]
data-art = "data_art.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/data_art"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py311"
line-length = 88
exclude = ["text/"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
